split up ChannelCredentials.cs
